各类ID由Long转换为Integer

+++ src/main/java/otakuplus/straybird/othellogame/applicationstates/ApplicationConnectState.java
@@ -26,11 +26,11 @@ public class ApplicationConnectState implements ApplicationState {
 
 	}
 
-	public void enterGameTable(Long gameTableId,Long seatId) {
+	public void enterGameTable(Integer gameTableId,Integer seatId) {
 
 	}
 
-    public void leaveGameTable(Long gameTableId,Long seatId){
+    public void leaveGameTable(Integer gameTableId,Integer seatId){
 
     }
