+++ src/main/java/otakuplus/straybird/othellogame/applicationstates/ApplicationIntializeState.java
@@ -37,10 +37,10 @@ public class ApplicationIntializeState implements ApplicationState {
 
 	}
 
-	public void enterGameTable(Long gameTableId,Long seatId) {
+	public void enterGameTable(Integer gameTableId,Integer seatId) {
 	}
 
-	public void leaveGameTable(Long gameTableId,Long seatId){
+	public void leaveGameTable(Integer gameTableId,Integer seatId){
 
 	}
 	public void logout(){
