+++ src/main/java/otakuplus/straybird/othellogame/applicationstates/ApplicationLeaveGameHallState.java
@@ -52,11 +52,11 @@ public class ApplicationLeaveGameHallState implements ApplicationState{
         }
     }
 
-    public void enterGameTable(Long gameTableId,Long seatId) {
+    public void enterGameTable(Integer gameTableId,Integer seatId) {
 
     }
 
-    public void leaveGameTable(Long gameTableId,Long seatId){
+    public void leaveGameTable(Integer gameTableId,Integer seatId){
 
     }
