+++ src/main/java/otakuplus/straybird/othellogame/applicationstates/ApplicationLogoutState.java
@@ -29,11 +29,11 @@ public class ApplicationLogoutState implements ApplicationState{
 
     }
 
-    public void enterGameTable(Long gameTableId,Long seatId) {
+    public void enterGameTable(Integer gameTableId,Integer seatId) {
 
     }
 
-    public void leaveGameTable(Long gameTableId,Long seatId){
+    public void leaveGameTable(Integer gameTableId,Integer seatId){
 
     }
